banking/requirements.txt
plaid-python>=10.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
requests>=2.28.0
